pyproject.toml
[tool.poetry]
name = "target-weaviate"
version = "0.0.1"
description = "`target-weaviate` is a Singer target for Weaviate, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Steven Searcy <steven@foodhealth.co>"]
keywords = [
    "ELT",
    "Weaviate",
    "Singer",
    "Meltano"
]
license = "Apache-2.0"
packages = [
    { include = "target_weaviate" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
singer-sdk = "~=0.39.0"
weaviate-client = "^4.9.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
singer-sdk = { version="~=0.39.0", extras = ["testing"] }
ruff = "^0.1.0"
poetry-dynamic-versioning = "^1.0.0"

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    "COM812",  # missing-trailing-comma
    "ISC001",  # single-line-implicit-string-concatenation
]
select = ["ALL"]
src = ["target_weaviate"]
target-version = "py38"

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["target_weaviate"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
# CLI declaration
target-weaviate = 'target_weaviate.target:TargetWeaviate.cli'

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

